Refactor raw_filters in categories definitions and allow new 'cuisine' filter
