+++ idunn/api/categories.py
@@ -5,17 +5,12 @@ from typing import List
 
 class CategoryDescription(BaseModel):
     name: str = Field(..., description="Unique label of the category.")
-    raw_filters: List[str] = Field(..., description="Raw filter on OSM tags.")
 
 
 class AllCategoriesResponse(BaseModel):
     categories: List[CategoryDescription] = Field(..., description="All available categories")
 
 
-def get_all_categories() -> AllCategoriesResponse:
+def get_all_categories():
     """List all available categories."""
-    return {
-        "categories": [
-            {"name": cat.value, "raw_filters": cat.raw_filters()} for cat in list(Category)
-        ]
-    }
+    return {"categories": [{"name": cat.value} for cat in list(Category)]}
