+++ idunn/api/utils.py
@@ -1,5 +1,4 @@
 from enum import Enum
-from fastapi import HTTPException
 from elasticsearch import (
     Elasticsearch,
     ConnectionError,
@@ -26,10 +25,8 @@ from idunn.blocks import (
     WikiUndefinedException,
 )
 from idunn.utils import prometheus
-from idunn.utils.index_names import INDICES
-from idunn.utils.es_wrapper import get_elasticsearch
-from idunn.places.exceptions import PlaceNotFound
 from idunn.utils.settings import _load_yaml_file
+from idunn.datasources.mimirsbrunn import MimirPoiFilter
 
 logger = logging.getLogger(__name__)
 
@@ -75,8 +72,15 @@ class CategoryEnum(str):
     def pj_what(self):
         return ALL_CATEGORIES[self].get("pj_what")
 
-    def raw_filters(self):
-        return ALL_CATEGORIES[self].get("raw_filters")
+    def raw_filters(self) -> [MimirPoiFilter]:
+        raw_filters = ALL_CATEGORIES[self].get("raw_filters")
+        filters = []
+        for f in raw_filters:
+            f = f.copy()
+            poi_class = f.pop("class", None)
+            poi_subclass = f.pop("subclass", None)
+            filters.append(MimirPoiFilter(poi_class, poi_subclass, extra=f))
+        return filters
 
     def regex(self):
         return ALL_CATEGORIES[self].get("regex")
@@ -149,13 +153,6 @@ BLOCKS_BY_VERBOSITY = {
     Verbosity.SHORT: [OpeningHourBlock, Covid19Block],
 }
 
-PLACE_DEFAULT_INDEX = settings["PLACE_DEFAULT_INDEX"]
-PLACE_POI_INDEX = settings["PLACE_POI_INDEX"]
-PLACE_ADDRESS_INDEX = settings["PLACE_ADDRESS_INDEX"]
-PLACE_STREET_INDEX = settings["PLACE_STREET_INDEX"]
-
-ANY = "*"
-
 
 class WikidataConnector:
     _wiki_es = None
@@ -215,153 +212,6 @@ class WikidataConnector:
 
         wiki = resp[0]["_source"]
         return wiki
-
-
-def fetch_es_poi(id, es) -> dict:
-    """Returns the raw POI data
-    @deprecated by fetch_es_place()
-
-    This function gets from Elasticsearch the
-    entry corresponding to the given id.
-    """
-    try:
-        return fetch_es_place(id, es, type="poi")["_source"]
-    except PlaceNotFound as e:
-        raise HTTPException(status_code=404, detail=e.message) from e
-
-
-def fetch_es_pois(raw_filters, bbox, max_size) -> list:
-    es = get_elasticsearch()
-    left, bot, right, top = bbox[0], bbox[1], bbox[2], bbox[3]
-
-    terms_filters = []
-    for pair in raw_filters:
-        (cls, subcls) = pair.split(",", 1)
-        if (cls, subcls) == (ANY, ANY):
-            terms_filters.append([])
-        elif subcls == ANY:
-            terms_filters.append(["class_" + cls])
-            terms_filters.append([cls])
-        elif cls == ANY:
-            terms_filters.append(["subclass_" + subcls])
-        else:
-            terms_filters.append(["class_" + cls, "subclass_" + subcls])
-
-    should_terms = []
-    for filt in terms_filters:
-        if filt == []:
-            should_terms.append({"match_all": {}})
-        else:
-            should_terms.append(
-                {"bool": {"must": [{"term": {"poi_type.name": term}} for term in filt]}}
-            )
-
-    # pylint: disable = unexpected-keyword-arg
-    bbox_places = es.search(
-        index=INDICES["poi"],
-        body={
-            "query": {
-                "bool": {
-                    "should": should_terms,
-                    "minimum_should_match": 1,
-                    "filter": {
-                        "geo_bounding_box": {
-                            "coord": {
-                                "top_left": {"lat": top, "lon": left},
-                                "bottom_right": {"lat": bot, "lon": right},
-                            }
-                        }
-                    },
-                }
-            },
-            "sort": {"weight": "desc"},
-        },
-        size=max_size,
-        timeout="3s",
-        ignore_unavailable=True,
-    )
-
-    bbox_places = bbox_places.get("hits", {}).get("hits", [])
-    return bbox_places
-
-
-def fetch_es_place(id, es, type) -> dict:
-    """Returns the raw Place data
-
-    This function gets from Elasticsearch the
-    entry corresponding to the given id.
-    """
-    if type is None:
-        index_name = PLACE_DEFAULT_INDEX
-    elif type not in INDICES:
-        raise HTTPException(status_code=400, detail=f"Wrong type parameter: type={type}")
-    else:
-        index_name = INDICES.get(type)
-
-    try:
-        es_places = es.search(
-            index=index_name,
-            body={"query": {"bool": {"filter": {"term": {"_id": id}}}}},
-            ignore_unavailable=True,
-            _source_exclude=["boundary"],
-        )
-    except ElasticsearchException as error:
-        logger.warning("error with database: %s", error)
-        raise HTTPException(detail="database issue", status_code=503) from error
-
-    es_place = es_places.get("hits", {}).get("hits", [])
-    if len(es_place) == 0:
-        if type is None:
-            message = f"place '{id}' not found"
-        else:
-            message = f"place '{id}' not found with type={type}"
-        raise PlaceNotFound(message=message)
-    if len(es_place) > 1:
-        logger.warning("Got multiple places with id %s", id)
-
-    return es_place[0]
-
-
-def fetch_closest(lat, lon, max_distance, es):
-    es_addrs = es.search(
-        index=",".join([PLACE_ADDRESS_INDEX, PLACE_STREET_INDEX]),
-        body={
-            "query": {
-                "function_score": {
-                    "query": {
-                        "bool": {
-                            "filter": {
-                                "geo_distance": {
-                                    "distance": "{}m".format(max_distance),
-                                    "coord": {"lat": lat, "lon": lon},
-                                    "distance_type": "plane",
-                                }
-                            }
-                        }
-                    },
-                    "boost_mode": "replace",
-                    "functions": [
-                        {
-                            "gauss": {
-                                "coord": {
-                                    "origin": {"lat": lat, "lon": lon},
-                                    "scale": "{}m".format(max_distance),
-                                }
-                            }
-                        }
-                    ],
-                }
-            },
-            "from": 0,
-            "size": 1,
-        },
-    )
-    es_addrs = es_addrs.get("hits", {}).get("hits", [])
-    if len(es_addrs) == 0:
-        raise HTTPException(
-            status_code=404, detail=f"nothing around {lat}:{lon} within {max_distance}m..."
-        )
-    return es_addrs[0]
 
 
 def build_blocks(es_poi, lang, verbosity):
