+++ tests/fixtures/cinema_multiplexe.json
@@ -105,7 +105,7 @@
   ],
   "poi_type": {
     "id": "cinema",
-    "name": "cinema"
+    "name": "class_cinema"
   },
   "properties": [
     {
