+++ tests/fixtures/fake_all_blocks.json
@@ -387,7 +387,7 @@
   "poi_type": [
     {
       "id": "museum",
-      "name": "museum"
+      "name": "class_museum"
     }
   ],
   "properties": [
