+++ tests/fixtures/patisserie_peron.json
@@ -105,7 +105,7 @@
   ],
   "poi_type": {
     "id": "bakery",
-    "name": "bakery"
+    "name": "class_bakery"
   },
   "properties": [
     {
